PlanRouter uses TripStructureUtils instead of TripRouter.tripsToLegs
Converting trips to legs becomes more what it should be: a hack used for
backward compatibility, not a part of the core behavior.

A more modular way to identify the main mode is introduced: rather than
requiring to extend the TripRouter, the main mode identification can be
done by providing a MainModeIdentifier object to the PlanRouter.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@22774 a0517364-e873-4167-b1c8-416daa5a6b91