buildPlanString now just returns the solution of the "subproblem" at hand
before, it returned a PlanString corresponding to the solution of the
full problem. Now, it returns a PlanAllocation corresponding only to the
subproblem (the problem of allocating plans for a given subset of the
agents), at least when blocking combinations are not forbidden (in which
case the subproblem also depends on the allocation "above").
This opens the way to caching the solution to subproblems, which
solutions were until now computed over and over.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@23435 a0517364-e873-4167-b1c8-416daa5a6b91