new selector
still untested

agents are parsed in random order. The best possible plan for each
agent, as well as the linked plans, are selected.
Hence, one agent (the boss), chooses the joint plan because it is good
for him, the others follow.

Idea to improve: restrict who can be the boss (eg. only drivers)

git-svn-id: file:///srv/local/data/cleanRepo/trunk@25755 a0517364-e873-4167-b1c8-416daa5a6b91