a performance test of the SPSC queue vs blocking queue
the SPSC queue is indeed faster, but I am absolutely sure it does not
matter at all. With as few as 4 event threads on a ZH 10% scenario, more
than 90% of the time of a thread is passed waiting for an event to
come.

One could try to transform the SPSC list into a blocking queue, but this
is rather tricky: just let things like this for the moment.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@28789 a0517364-e873-4167-b1c8-416daa5a6b91